yoga-pose/requirements.txt
pdf2image
pdfminer
unstructured
pillow-heif
opencv-python-headless
pikepdf 
pypdf
google-generativeai
